requirements.txt
google_api_python_client==2.68.0
google_auth_oauthlib==0.7.0
pandas==1.4.4
protobuf==4.21.10
PyDrive==1.3.1
streamlit==1.12.2
